requirements.txt
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pygsheets==2.0.3.1
pyOpenSSL==19.1.0
python-pcre==0.7
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
